refactor(api): colocate auth seeding helpers
