+++ apps/api/src/modules/auth/auth.seeder.ts
@@ -1,8 +1,7 @@
 import { faker } from '@faker-js/faker'
 import { Dictionary, EntityManager } from '@mikro-orm/core'
 import { Seeder } from '@mikro-orm/seeder'
-import { hashPassword } from 'better-auth/crypto'
-import { Account, User } from '../modules/auth/auth.entity'
+import { createUserData } from './auth.factory'
 
 const password = 'Password123!'
 
@@ -22,23 +21,12 @@ export class AuthSeeder extends Seeder {
   async run(em: EntityManager, context: Dictionary): Promise<void> {
     context.users = []
     for (const userData of [defaultUser, ...users]) {
-      // Create user
-      const user = new User()
-      user.name = userData.name
-      user.email = userData.email.toLowerCase()
-      user.emailVerified = true
-      await em.persist(user).flush()
+      const user = await createUserData(em, {
+        name: userData.name,
+        email: userData.email.toLowerCase(),
+        emailVerified: true,
+      }, userData.password)
       context.users.push(user)
-
-      // Create account with password
-      const account = new Account()
-      account.user = user
-      account.providerId = 'credential'
-      account.accountId = crypto.randomUUID()
-      // Store the password directly
-      account.password = await hashPassword(userData.password)
-
-      await em.persist(account).flush()
     }
   }
 }
